Add `--hspec` option to generate HSpec specs
Previously, `huspenders` only generated an executable, that is, something you
can run with `cabal run`.

But HSpec needs to be able to import your code, and to expose your code, it must
be in a library.

So now `huspenders` generates both a library and an executable:

* `src/MyCoolProject.hs` should be where your "real" code goes
* `src/Main.hs` imports `src/MyCoolProject.hs`, so the library code can be used
  by Main for `cabal run`

Additionally, the script generates less output and relies on the final green
message to reassure users that everything worked.