Don't initialize git if we're already in a repo
If we're already in a git repo, initializing a new repo _probably_ isn't
what you want to do. If it is, I think it's fair to assume you'll do it
manually.

To that end, I think it's reasonable to only run `git init` if we're
currently _not_ in a git repo.